Adds about route in client
