+++ services/client/src/components/About.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const About = () => {
+  return (
+    <div>
+      <h1 className="title is-1">About</h1>
+      <hr />
+      <br />
+      <p className="content">About section</p>
+    </div>
+  );
+};
+
+export default About;
