+++ services/client/src/components/__tests__/About.test.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { render, cleanup } from "@testing-library/react";
+
+import About from "../About";
+
+afterEach(cleanup);
+
+it("renders about", () => {
+  const { getByText } = render(<About />);
+  expect(getByText("About section")).toHaveClass("content");
+});
+
+it("renders", () => {
+  const { asFragment } = render(<About />);
+  expect(asFragment()).toMatchSnapshot();
+});
