+++ services/client/src/components/__tests__/App.test.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { render, cleanup } from "@testing-library/react";
+
+import App from "../../App";
+
+afterEach(cleanup);
+
+it("renders", () => {
+  const { asFragment } = renderWithRouter(<App />);
+  expect(asFragment()).toMatchSnapshot();
+});
