+++ services/client/src/setupTests.js
@@ -2,4 +2,21 @@
 // allows you to do things like:
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
+import React from "react";
+import { Router } from "react-router-dom";
+import { render } from "@testing-library/react";
+import { createMemoryHistory } from "history";
 import "@testing-library/jest-dom/extend-expect";
+
+global.renderWithRouter = function renderWithRouter(
+  ui,
+  {
+    route = "/",
+    history = createMemoryHistory({ initialEntries: [route] }),
+  } = {}
+) {
+  return {
+    ...render(<Router history={history}>{ui}</Router>),
+    history,
+  };
+};
